Move function as static helpers on the model
