+++ lib/helper/premade.js
@@ -1,8 +1,5 @@
 "use strict";
 
-var mongoose = require("mongoose");
-var rarity = require("rarity");
-
 
 /**
  * Recursive function for premade,
@@ -87,54 +84,4 @@ module.exports.getPremade = function(premadeData) {
   });
 
   return premade;
-};
-
-
-// Some premades are retrieved from match history, but for better accuracy most of them are stored in the DB.
-// This function retrieves all relevant documents for specified summonerIds
-module.exports.loadKnownPremadesFromDB = function loadKnownPremadesFromDB(summonerIds, region, cb) {
-  var prefix = region.toLowerCase() + ":";
-  var lookingFor = summonerIds.map(function(id) {
-    return prefix + id;
-  });
-
-  mongoose.model("Premade").find({_id: {$in: lookingFor}}, function(err, premades) {
-    if(err) {
-      return cb(err);
-    }
-
-    var ret = {};
-    premades.forEach(function(premade) {
-      ret[premade._id.replace(prefix, '')] = premade.premades.toObject();
-    });
-
-    cb(null, ret);
-  });
-};
-
-
-// Save new premades data in an atomic way to mongo
-// premades should be the output of the getPremades() function
-module.exports.savePremadesToDB = function savePremadesToDB(premadesObject, region, cb) {
-  var prefix = region.toLowerCase() + ":";
-  var premades = Object.keys(premadesObject).reduce(function(acc, key) {
-    return acc.concat(premadesObject[key]);
-  }, []);
-
-  var bulk = mongoose.model('Premade').collection.initializeUnorderedBulkOp();
-  premades.forEach(function(premade) {
-    premade.forEach(function(summonerId) {
-      bulk.find({_id: prefix + summonerId}).upsert().updateOne({
-        $addToSet: {
-          premades: {
-            $each: premade.filter(function(id) {
-              return id !== summonerId;
-            })
-          }
-        }
-      });
-    });
-  });
-
-  bulk.execute(rarity.slice(1, cb));
 };
